backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
stripe==10.12.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
email-validator==2.2.0
